refactor: repository제거
