+++ app/src/main/java/woowacourse/movie/UserTicketApplication.kt
@@ -1,11 +1,11 @@
 package woowacourse.movie
 
 import android.app.Application
-import woowacourse.movie.model.db.UserTicketRepository
+import woowacourse.movie.model.db.UserTicketDatabase
 
 class UserTicketApplication : Application() {
     override fun onCreate() {
         super.onCreate()
-        UserTicketRepository.initialize(this)
+        UserTicketDatabase.initialize(this)
     }
 }
