+++ app/src/main/java/woowacourse/movie/ui/booking/MovieBookingHistoryPresenter.kt
@@ -2,17 +2,17 @@ package woowacourse.movie.ui.booking
 
 import android.os.Handler
 import android.os.Looper
-import woowacourse.movie.model.db.UserTicketRepository
+import woowacourse.movie.model.db.UserTicketDatabase
 
 class MovieBookingHistoryPresenter(
     private val view: MovieBookingHistoryContract.View,
 ) :
     MovieBookingHistoryContract.Presenter {
     override fun loadBookingHistories() {
-        val userTicketRepository = UserTicketRepository.instance()
+        val db = UserTicketDatabase.database()
         val handler = Handler(Looper.getMainLooper())
         Thread {
-            val userTickets = userTicketRepository.getUserTickets()
+            val userTickets = db.userTicketDao().findAll()
             handler.post {
                 view.showBookingHistories(userTickets)
             }
